+++ pytest_helm_charts/giantswarm_app_platform/app.py
@@ -20,8 +20,6 @@ class AppFactoryFunc(Protocol):
         namespace: str = "default",
         deployment_namespace: str = "default",
         config_values: YamlDict = None,
-        namespace_config_annotations: YamlDict = None,
-        namespace_config_labels: YamlDict = None,
         extra_metadata: Optional[dict] = None,
         extra_spec: Optional[dict] = None,
         timeout_sec: int = 60,
@@ -44,8 +42,6 @@ def app_factory_func(
         namespace: str = "default",
         deployment_namespace: str = "default",
         config_values: YamlDict = None,
-        namespace_config_annotations: YamlDict = None,
-        namespace_config_labels: YamlDict = None,
         extra_metadata: Optional[dict] = None,
         extra_spec: Optional[dict] = None,
         timeout_sec: int = 60,
@@ -66,10 +62,6 @@ def app_factory_func(
              deployment_namespace: namespace where the app will be deployed (can be different than `namespace`)
              config_values: any values that should be used to configure the app (same as `values.yaml` used for
                 a Helm Chart directly).
-             namespace_config_annotations: a dictionary of annotations that need to be added to the
-                `deployment_namespace` created for the app
-             namespace_config_labels: a dictionary of labels that need to be added to the `deployment_namespace`
-                created for the app
              extra_metadata: optional dict that will be merged with the 'metadata:' section of the object
              extra_spec: optional dict that will be merged with the 'spec:' section of the object
              timeout_sec: timeout in seconds for the create operation
@@ -95,8 +87,6 @@ def app_factory_func(
             namespace,
             deployment_namespace,
             config_values,
-            namespace_config_annotations,
-            namespace_config_labels,
             extra_metadata,
             extra_spec,
         )