+++ pytest_helm_charts/giantswarm_app_platform/utils.py
@@ -117,8 +117,6 @@ def make_app_object(
     namespace: str,
     deployment_namespace: str,
     config_values: YamlDict = None,
-    namespace_config_annotations: YamlDict = None,
-    namespace_config_labels: YamlDict = None,
     extra_metadata: Optional[dict] = None,
     extra_spec: Optional[dict] = None,
 ) -> ConfiguredApp:
@@ -137,10 +135,6 @@ def make_app_object(
         deployment_namespace: namespace where the app will be deployed (can be different than `namespace`)
         config_values: any values that should be used to configure the app (same as `values.yaml` used for
             a Helm Chart directly).
-        namespace_config_annotations: a dictionary of annotations that need to be added to the
-            `deployment_namespace` created for the app
-        namespace_config_labels: a dictionary of labels that need to be added to the `deployment_namespace`
-            created for the app
         extra_metadata: optional dict that will be merged with the 'metadata:' section of the object
         extra_spec: optional dict that will be merged with the 'spec:' section of the object
 
@@ -149,10 +143,6 @@ def make_app_object(
     """
     if config_values is None:
         config_values = {}
-    if namespace_config_annotations is None:
-        namespace_config_annotations = {}
-    if namespace_config_labels is None:
-        namespace_config_labels = {}
 
     # TODO: include proper regexp validation
     assert app_name != ""
@@ -176,10 +166,6 @@ def make_app_object(
             "kubeConfig": {"inCluster": True},
             "name": app_name,
             "namespace": deployment_namespace,
-            "namespaceConfig": {
-                "annotations": namespace_config_annotations,
-                "labels": namespace_config_labels,
-            },
         },
     }
     app_cm_obj: Optional[ConfigMap] = None
@@ -206,8 +192,6 @@ def create_app(
     namespace: str,
     deployment_namespace: str,
     config_values: YamlDict = None,
-    namespace_config_annotations: YamlDict = None,
-    namespace_config_labels: YamlDict = None,
     extra_metadata: Optional[dict] = None,
     extra_spec: Optional[dict] = None,
 ) -> ConfiguredApp:
@@ -220,8 +204,6 @@ def create_app(
         namespace,
         deployment_namespace,
         config_values,
-        namespace_config_annotations,
-        namespace_config_labels,
         extra_metadata,
         extra_spec,
     )