+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), following [Se
   - cleanups:
     - all methods creating objects, but not submitting them to k8s API, have now `make-*-object` pattern
     - `HTTPClient` connection object is always the first parameter
+    - removed the `namespaceConfig*` arguments of `make_app_object`, as they can be now included
+      using the `extra-*` args
 
 ## [0.6.0]